pyproject.toml
[project]
name = "anu-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "langchain-google-genai>=4.1.2",
    "langchain[google-genai]>=1.2.3",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-cli[inmem]>=0.4.4",
    "python-dotenv>=1.1.1",
    "sse-starlette>=2.1.3",
    "tavily-python>=0.7.17",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
